# swagger/v1/swagger.yaml
...
openapi: 3.0.1
info:
  title: ProxySock API
  version: v1
  description: API for Resellers to manage proxies, VMs, and orders.
paths:
...
  "/api/v1/resellers/{id}":
    parameters:
    - name: id
      in: path
      required: true
      schema:
        type: string
    get:
      summary: Get Reseller Profile
      tags:
      - Reseller Profile
      security:
      - Bearer: []
      responses:
        '200':
          description: successful
    patch:
      summary: Update Reseller Profile
      tags:
      - Reseller Profile
      security:
      - Bearer: []
      parameters: []
      responses:
        '200':
          description: successful
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                company_name:
                  type: string
                  example: Updated Partner Corp
                email:
                  type: string
                  example: new@example.com
  "/api/v1/resellers":
    get:
      summary: List Resellers
      tags:
      - Reseller Profile
      security:
      - Bearer: []
      description: Returns a paginated list of authorized partners.
      responses:
        '200':
          description: successful
...
servers:
- url: https://{defaultHost}
  variables:
    defaultHost:
      default: api.proxysock.com
- url: http://localhost:3000
  description: Local development server
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your JWT token here. Get it from /api/v1/auth/token
security:
- Bearer: []